move GetColorPtr
move GetColorPtr
